fix: move export logic to replicator
The logic changed in the last rebase, which wasn't taken into account.
This should make us use a single worker again.
