+++ pkg/registry/apis/provisioning/register.go
@@ -186,12 +186,6 @@ func (b *ProvisioningAPIBuilder) UpdateAPIGroupInfo(apiGroupInfo *genericapiserv
 		b.logger.With("worker", "github"),
 		provisioning.IncludeYamlOrJSON,
 	))
-	b.jobs.Register(&exportWorker{
-		getter:         b,
-		logger:         b.logger.With("worker", "export"),
-		resourceClient: b.client,
-		identities:     b.identities,
-	})
 
 	repositoryStorage.AfterCreate = b.afterCreate
 	// AfterUpdate doesn't have the old object, so we have to use BeginUpdate
