+++ pkg/registry/apis/provisioning/resources/replicator_test.go
@@ -1,4 +1,4 @@
-package provisioning
+package resources
 
 import (
 	"testing"
